mtd: fixing CFI for Numonyx CFI bug
commit 6a7bb4c6969e2b3e36796d649609216b3deffa40 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

The M29EW devices seem to report the CFI information wrong when
it's in 8 bit mode.

There's an app note from Numonyx on this issue and there's a patch
in the open source, but it doesn't seem to be in the mainline.

This commit is based on that patch from Massimo Cirillo titled:
buffer size adjustment for M29EW Numonyx devices.

Signed-off-by: John Linn <john.linn@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>