drm: Add adv7511 encoder driver
commit  adba6aa6d084ae91b655d59a47f33daf1d328eeb from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

This patch adds a driver for the Analog Devices adv7511. The adv7511 is
a standalone HDMI transmitter chip. It features a HDMI output interface
on one end and video and audio input interfaces on the other.

To avoid the conflicts with drm bridge adv7511 driver,
add _LEGACY at the end of the identifier string. This config
depends on !DRM_I2C_ADV7511 so that only one of them can be
enbled at a time. While at it, keep 'depends on OF'.

This is a temporary solution to enable adv7511 based on the encoder
slave. The DRM driver needs to support the atomic modesetting
as well as the drm bridge. When those changes are in, this can be
removed.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>
Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>