zynq: Add OCM driver
commit f2f377be5396fa13f6ecfc19d68b28047211c43d from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

The driver provides memory allocator which can
be used by others drivers to allocate memory inside OCM.
All location for 64kB blocks are supported
and driver is trying to allocate the largest continuous
block of memory.

Checking mpcore addressing filterring is not done here
but could be added in future.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>