dts: s32v: DRAM correct size of DDR regions in dts and dtsi files
commit 0f3a29ef416a2a70e57a98f0fc5744553f3a5cf1 from
https://source.codeaurora.org/external/autobsps32/linux release/bsp31.0_cd-5.10.41-rt

The size of DDR regions was corrected in dts files (dtsi file of evb,
because this is used by more evb dts files). In the SoC dtsi file,
it is used the size of DDR than can be mapped.
The board dts files will then overwrite these settings with their
available DDR size.

Issue: ALB-5945

Signed-off-by: Ana Nedelcu <ana.nedelcu@nxp.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>