media: xilinx: dma: Enable use of Framebuffer Write IP in V4L2 Driver
commit fa39d1e7a749efbc7aa8aec6d2411ddbc4f0622a from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

The Xilinx Video Framebuffer driver is video 'format aware' and
requires clients to pass video format information as part of DMA
programming. As the DMA Engine interface does not support passing
this information, a configuration object is passed through the
dma channel's private data to the Framebuffer Driver with this
information.

Signed-off-by: Jeffrey Mouroux <jeff.mouroux@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>