media: xilinx: Ensure Video Framebuffer selected by Kconfig
commit  e60fb63d59cdd7752c300a3c0966452df90aa0c0 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

The Xilinx Video Linux driver is a client of the Video Framebuffer DMA
driver. Because the Framebuffer driver exports symbols used by this driver,
it is important that both are built in the same manner: as modules or
as statically included drivers.

Signed-off-by: Jeffrey Mouroux <jeff.mouroux@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>