uvcvideo: Prevent new URBs being processed at stream stop
commit  7d6a12ac4b8941ff8feb53be5b66c6e0bfbe1f52 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

With asynchronous handling of the URBs from the USB Complete handler, we
get a continual stream of packets being received while we are attempting
to shutdown the stream.

Packets that have already been received and processed are on a
work-queue, but during stream shutdown the URBs that those packets
belong to are killed and free'd.

To prevent this race from causing invalid memory accesses, prevent new
URBs from being processed when uvc_stop_streaming() is called by
introducing a new flag "UVC_QUEUE_STOPPING" into the queue, and checking
this when processing the URB to be queued.

With this, we can flush the work queue, and commence a normal pipe
shutdown. Work tasks that are already queued are processed, but the URBs
are prevented from being re-submitted to the USB stack.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Christian Kohn <christian.kohn@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
[Quanyang: use dev_err instead of uvc_printk because of mainline commit 69df09547e7a
("media: uvcvideo: Use dev_ printk aliases").]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>