drm: xlnx: fb: Prefer the current format depth over depth from fb helper
commit  72a0a7a136d453149dccd892ebabcb838028bd77 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

The drm fb helper has specific preference of bpp and depth. For example,
for 32bit bpp, the depth is hard-coded to be 24. If it's not aligned
with the supported format of a drm device, it fails to initialize fbdev.
So override the depth value from fb helper with the current format
of the drm device. This will allow to initialize the fbdev with
preferred format that matches with actual format.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Reviewed-by: Saurabh Sengar <saurabhs@xilinx.com>
Tested-by: Anil Kumar Mamidala <amamidal@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>