drm: xlnx: zynqmp_disp: Modul param for initial format of gfx layer
commit  2ea6f218fad8b406f57d576efe7ccb3438e116b3 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.4

This allows the initial format for the graphics layer to be specified
and changed through module param. The format is used for
the fbdev emulation, and the emulated fbdev will be initialized with
the specified format: 0 = rgb565, 1 = rgb888, 2 = argb8888.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Reviewed-by: Satish Kumar Nagireddy <satish.nagireddy.nagireddy@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>