drm: xlnx: sdi: Add support for Field 1 Active size
commit  718fc8663fece370a82971631071909168e60b12 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.4

Add support for the new Field 1 Active register in the VTC for
interlaced mode. In some interlaced modes, the field 0 and 1 active size
may be different. To support such modes, this new register is added to
the VTC. Writing to this register in progressive mode is optional.

Signed-off-by: Vishal Sagar <vishal.sagar@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>