xilinx: v4l: dma: Add format enumeration for output
commit  6229ede247f00cd016a8e9eda3c5ef66b90d13fc from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

This patch modifies the implementation of format enumeration to make
it generic. Logic checks if DMA is frmbuf or not and based on that it
invoke frmbuf specific APIs or generic APIs to enumerate the formats.

Signed-off-by: Satish Kumar Nagireddy <satish.nagireddy.nagireddy@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>