GPIO: Add GPIO driver to S32-gen1
commit 8205ee887d4a85425aba3c532b5baebabeb413ac from
https://source.codeaurora.org/external/autobsps32/linux release/bsp31.0_cd-5.10.41-rt

Add GPIO driver for S32-Gen1 to allow multiple instances of the driver
to share the same register ranges due to the non-contiguous
arrangement of the EIRQ pins. The device tree binding of
this driver will be changed to use the regmap API to allow
this sharing of resources.

Issue: ALB-3170
Issue: ALB-3677
Issue: ALB-4883
Issue: ALB-5974
Signed-off-by: Matthew Nunez <matthew.nunez@nxp.com>
Signed-off-by: Phu Luu An <phu.luuan@nxp.com>
Signed-off-by: Costea Ciprian <ciprianmarian.costea@nxp.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>