staging: xroeframer: add sysfs entries for "stats" registers
commit  6d6ecece2005ebfaf10352d42b96569d1680a7bc from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

This includes the first 2 registers of the stats register bank as sysfs
entries. The rest of the entries will be added on the next patch
and then the "/dev/xroefram/stats" device will be removed from
the driver's code.

Signed-off-by: Vasileios Bimpikas <vasileios.bimpikas@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>