media: xilinx: multi-scaler: Add extra padding bytes after each row of pixels
commit  f97942e72e65d496b8d17d73bfec2e06d0045609 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.4

Padding bytes is necessary to ensure that every row of pixels starts at
an address that is aligned with the size of the data on the memory
mapped interface. Therefore, the stride value must be a multiple of
the memory mapped data size. The data size of the memory mapped
interface is 64*Samples per Clock bits, for example 64, 128, or 256
bits for 1, 2, and 4 pixels per clock, respectively.

By introducing these changes, the backward compatibility is broken
between DT and driver. So, it is mandatory to take new DT with latest
driver.

Signed-off-by: Suresh Gupta <suresh.gupta@xilinx.com>
Reviewed-by: Satish Kumar Nagireddy <satish.nagireddy.nagireddy@xilinx.com>
Reviewed-by: Vishal Sagar <vishal.sagar@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>