Revert "Revert "dma: xilinx: Release buffers before DMA transfer""
commit  b3478d96e6443117d3398d3e249356e30c2a343e from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

We need the early callback for low latency to return buffer at
start of descriptor processing as without it there is no way for
dma client and hence also the application to be aware of
actual position of dma.

So without earlycallback, it leads to a problem as buffers
carry the timestamps based upon when they were queued and not
when dma has actually started writing onto it.

This reverts commit 4036801ee0c1ceef86bbe8dd2f3a39408b224161.

Signed-off-by: Devarsh Thakkar <devarsh.thakkar@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>