dma: xilinx: Disable low latency capture in file ops open
commit  56b1c8121b1de35a1ba3c80e2f914afbd3194c6d from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.4

Add a custom handler to be called when file handle is opened
by application to disable the low latency capture and reset
dmaengine to AUTO_RESTART mode. We do this if the file handle
is the only file handle instance opened for the associated
video device.

This is required in the scenario where application got
hung or crash or is killed using SIGKILL and is not able to
disable the low latency capture by itself. After this scenario,
the next run was getting affected where even if low latency
capture was not set, driver was still running in low latency
mode due to previous context which was not cleared.

Signed-off-by: Devarsh Thakkar <devarsh.thakkar@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Tested-by: Dylan Yip <dylan.yip@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>