v4l: xilinx: m2m: Fix crop support on capture port
commit  f3e984453744618a312f6791a3d9d202ececd52d from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

The driver has the implementation to support crop rectangle only on
the v4l2 output port (input side) and missing the support on v4l2
capture port (output side). Because of this, there is a hang observed
when application is requesting crop rectangle on the capture port.

This patch ensures that the DMA transfer is issued correctly when crop
rectangle is requested on the capture port.

Signed-off-by: Satish Kumar Nagireddy <satish.nagireddy.nagireddy@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>