arm: zynq: delete AFLAGS_suspend.o to fix compile warning
commit 8b50e1645cb201c94e5bd739c873114a1b381812 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

When using arm-linux-gnueabihf-gcc v9.2.0 to compile kernel, there will be
warning info as below:

cc1: warning: switch '-mcpu=cortex-a9' conflicts with '-march=armv7-a' switch

This is because that if "-march" and "-mcpu" options are passed to gcc together,
gcc will check if there is conflict between the two options unconsidering "quirk"
and "fpu" features. In gcc, the commit e319ca91f665 ("PR target/88799 Add +mp
and +sec extensions to ARMv7-a") add "mp" and "sec" features to "cortex-a9".
So the check will fail because the features in "cortex-a9" and "armv7-a" are no
longer the same.

Since suspend.S doesn't contain any "mp (Multiprocessing Extensions for v7-A
and v7-R architectures)" and "sec" (Security Extensions for v6K and v7-A
architectures) instruction, and "-march=armv7-a" has been defined in arch/arm/Makefile,
the "-mcpu=cortex-a9" is redudant here.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>