v4l: xilinx: sdirxss: Add YUV444 10 bpc support
commit  95d145051617ad41f3597c647fa92bec0c996193 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.4

This patch enables the YUV444 format in the Native SDI bridge when the
color format in the received ST352 packet is YUV 444.

Setting this bit in case of YUV422/420 source and keeping this bit
cleared in case of YUV444 source will lead to video corruption.

Signed-off-by: Vishal Sagar <vishal.sagar@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>