tty: pl011: Add support for configurable wordlength
commit  ee607d64ebbccbeada08565c31d89e6586467487 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.4

Currently we are not writing to the line control register
to update the wordlength.
Add support for configurable word length by updating the lcr.

Currently if we echo characters to the non-console uart the
no characters are displayed. The patch fixes the same.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>