dt-bindings: Add binding for IDT 8T49N24x UFT
commit 8fd3449abac6a7aaded32a953747cf3d4443f122 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

IDT8T49N241 has 4 outputs; 1 integral divider and 3 fractional dividers.
The 8T49N241 accepts up to two differential or single-ended input clocks
and a fundamental-mode crystal input. The internal PLL can lock to either
of the input reference clocks or just to the crystal to behave as a
frequency synthesizer.

Signed-off-by: David Cater <david.cater@idt.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>