Revert "net: phy: ensure phylib state machine is stopped after calling phy_stop"
commit 9b7124fb6b3b72401ad909e6cb00217140cf873d from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

This reverts commit cbfd12b3e8c3542e8142aa041714ed614d3f67b0.

An issue was observed with TI DP83867 PHYs on Xilixn boards where ifconfig
down/up does not work. This was traced to changes in phy_stop where the
state machine is stopped. It is as yet unclear why this is affecting this
PHY and it is being debugged. These patches are being reverted as a
workaround because this affects basic functionality.

Signed-off-by: Harini Katakam <harini.katakam@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>