mtd: rawnand: pl353: Fix column and row address calculation
commit  5e71c3be8993a169671f5745898aaa1a219fad47 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

The existing column address claculation is ignoring the 5th address cycle.
Ex: if flash address requested is 0x00001ffc, then the column is 0x1 and
row address should be 0xffc00000 which is totally of 5 address cycles.
but the existing implementation is splitting the column address as 0x1
and row address as 0x00000ffc which is incorrect.
update it properly inorder to read bbt correctly.

Signed-off-by: Naga Sureshkumar Relli <naga.sureshkumar.relli@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>