drm/fb_helper: Enable drm_leak_fbdev_smem by default for zynqmp
commit 306c2e38077ce19ab0ffad67093913ac79eaa4ef from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

As Mali400 user-space implementation relies on the physical address to
render into framebufffer and lot of use-case depends on this backend,
enable the module parameter 'drm_leak_fbdev_smem' by default.

The feature is deprecated in upstream due to the security concern, but
no alternative solution is given. For more information, refer to kernel
commit 4be9bd10e22d ("drm/fb_helper: Allow leaking fbdev smem_start").

This patch is Xilinx specific and not upstreamable.

Signed-off-by: Madhurkiran Harikrishnan <madhurkiran.harikrishnan@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>