i2c: xiic: Enter standard mode only for > 255 byte read transfers
commit  9c8ffb344015334cfc1b119a774e469bdeca583b from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

When standard mode is used for all the read transfers, timing of
interrupts is not in sync with the hardware. So, to avoid breaking
existing scenarios, enter standard mode only when the size of read
transfer is > 255 bytes.

Signed-off-by: Raviteja Narayanam <raviteja.narayanam@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>