v4l: xilinx: sdirxss: Fix returning error on stream off
commit  18fa67a01e060e080cfd3b1dc9821ca9e38ab917 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.4

When yavta app is started and the cable removed, a video unlock event
occurs. In this context the streaming is stopped as the bridges are
disabled. Now killing the yavta application calls s_stream() with enable
as false which returns -EINVAL.
So when yavta is restarted after cable is connected and video is locked,
the V4L framework doesn't call s_stream() with enable = true as it still
believes that SDI Rx subdev is still streaming.

This is fixed by returning success for s_stream() with enable = false
even when streaming is disabled.

Signed-off-by: Vishal Sagar <vishal.sagar@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>