v4l: xilinx: vipp: do enabling in the dependency check
commit  6565ccbb4223b71d09ceebf32746e652a8cd6b77 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.4

This patch moves the enabling/disabling as a part of dependency check.
Since it's not just a dependency check anymore, rename the function.
It removes the error return, so there can be false postive. Not fixing
here in this commit because it'll be removed in following patch anyway.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Reviewed-by: Dylan Yip <dylan.yip@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>