v4l: xilinx: sdirxss: Fix IP reset on gt clock switch in 12G mode
commit  df19308da4ce793cfe401c0bdd3b5dfe58945feb from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.4

On changing the GT clock from integer to fractional or vice versa, it is
observed that the IP resets. So the core is unable to detect the
incoming stream. The IP needs to be restored back to its state. So save
the control values and use them to reconfgure the SDI Rx IP after the GT
clock has changed.

Signed-off-by: Vishal Sagar <vishal.sagar@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>