usb: dwc3: of-simple: Enable CCI support in dwc3 driver
commit  8110b9b0dcabf1f2fcfcf8279e4aeec1f03bafcc from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.4

This patch adds CCI support in DWC3 driver when CCI is enabled in the
design. There are couple of registers which need to be modified for
making CCI enabled transfers in USB. This patch add the support for the
same.

Signed-off-by: Piyush Mehta <piyush.mehta@xilinx.com>
Signed-off-by: Manish Narani <manish.narani@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>