staging: apf: Make the APF config as a boolean state
commit  8df5cea3b28f4bad9695417289e65acbbbfbcd25 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.4

The xlnx.c uses low level functions such as __dma_map_area, that are
not exposed symbols. So it can't build as a separate kernel module,
but should be built as a kernel. So make the config as a boolean.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Reported-by: Florian Klink <flokli@flokli.de>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>