v4l: xilinx: dma: Add control handler to v4l2 device
commit 70272e385af589c70a2d4c18042db62aaa95be68 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Currently the control handler is not being passed to the v4l2 device. It
is only being held by video device. This causes a bug where the video
node is not aware of the controls implemented by its subdevices.

This patch passes the control handler to the v4l2 device.

Signed-off-by: Dylan Yip <dylan.yip@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>