dt-binding: soc: xilinx: ai-engine: Include SHIM row in partition
commit  f0bf45dcde54738933c851803efe45ee99441d63 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.4

Specify clearly that the SHIM row needs to be included in the partition
"reg" property. The start row "0" is already shown in the example, need to
add the explanation.

Also fix the number of rows to include the SHIM row in the example.

Signed-off-by: Wendy Liang <wendy.liang@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>