dt-binding: soc: xilinx: ai-engine: Include SHIM row in partition
commit  1ae8d861c9996fc6c6290c37101a602f044aea5d from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

Specify clearly that the SHIM row needs to be included in the partition
"reg" property. The start row "0" is already shown in the example, need to
add the explanation.

Also fix the number of rows to include the SHIM row in the example.

Signed-off-by: Wendy Liang <wendy.liang@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>