misc: xilinx-ai-engine: Update AI engine partition request parameter
commit  49fa146db4f5d380f2f6f2a3962868cba391e9e6 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.4

Update the parameters to request for an AI engine partition.
There are more that just partition ID should be used to request an AI
engine partition. Driver should check partition ID and also the UID
of the image loaded on the AI engine to see of it can grant the
application access to the partition. The request can also has metadata
for application to tell driver which resources the application uses. And
the application can use flag to tell the driver which particular
requirement the application needs, such as donot reset the partition
when closing the partition.

This commit is to update the parameters so that the user API doesn't
need to change in future when the underline feature is implemented.

Signed-off-by: Wendy Liang <wendy.liang@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>