misc: trafgen: fix out of scope warning
commit  8522fc547d2c4cf68ec6658b309510618e34c267 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

In xtg_pram_write the param_word is copied to data
and data is used after the scope of param_word. Fix the same
by moving the declaration of param_word. While at it more the
declaration of other variable to the start of the function.

Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>