ptp: Add Xilinx PTP timer driver
commit 98c6cc90c5d398e87b8602a47eb26f9702064a0d from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Add Xilinx PTP timer driver to be used with Xilinx MRMAC Ethernet
driver. This IP provides a PTP timer with support for time and fine
adjustment. It consists of a system timer and port timer (which is
used by the Ethernet MAC for timestamping). The port timer is used
for all time reporting and adjustment.

The timer clock period is design dependent but not currently available
via HW parameters. This upgrade will be added on top based on the
system design and hence this is marked as TODO item in the driver.

Signed-off-by: Harini Katakam <harini.katakam@xilinx.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>