misc: xilinx-ai-engine: Handle error interrupts for partitions not yet requested
commit  c864637eec95dd5a0988fd77da3d2b52066356ce from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

Error interrupts might be generated for partitions which are not yet
requested. Such errors could only be backtracked to the source module
when the tiles in-use are known. This change records the occurrence of
error interrupt for such a partition, and later schedules a task in
the workqueue when the partition is requested.

Signed-off-by: Nishad Saraf <nishad.saraf@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>