misc: xilinx-ai-engine: Clear broadcast event status
commit  16049279bd81db28a8be474faf295665f9baf881 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

The event broadcast from a tile above is logged in the event status
registers. As this status bit is used to determine if backtracking
the tile above is required, it needs to be cleared. Event status
bits are sticky and could be cleared by writing a 1.

Signed-off-by: Nishad Saraf <nishad.saraf@xilinx.com>
Reviewed-by: Wendy Liang <wendy.liang@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>