staging: xlnxsync: Reset channel error status after it is read
commit  5e4e1e46d74b25f2233b37cedbe73d17d7fedcce from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

Clear channel error status after it is read by the application.
The driver will not wait for application to clear it, instead once
it is read by application, it will reset it and then try to capture
new error status if any for subsequent frames and will be sending the
new status to application in the next channel status read.

The application needs to maintain the error status in it's context
for all the unresolved errors until they are resolved or become obsolete.

Signed-off-by: Devarsh Thakkar <devarsh.thakkar@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>