firmware: xilinx: Add shutdown/wakeup APIs
commit 867057fa3677ad3435e8f101cc13b0a4d3406250 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

Add shutdown/wakeup a resource eemi operations to shutdown
or bringup a resource.

Note alignment of args matches convention of other fn's in this file.
The reason being that the long fn name results in aligned args that
otherwise go over 80 chars so shift right to avoid this

Signed-off-by: Ben Levinsky <ben.levinsky@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Rajan Vaja <rajan.vaja@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
Link: https://lore.kernel.org/r/20201130155717.26008-3-ben.levinsky@xilinx.com
State: waiting (remoteproc)
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>