spi: spi-zynqmp-gqspi: Update correct value for DLY1 bit field
commit  89dd44831ba368b8386229d6241efab78fdbaba3 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

The driver was using an incorrect value for DLY1 bit field in the GQSPI
loopback clock delay adjustment register. This bit gets set in Versal
when the qspi operating frequency is greater than 100Mhz and the mode
was not validated earlier. Fixed it by updating DLY1 bit field in the
GQSPI loopback clock delay adjustment register with the correct value
as recommended in the IP spec.

Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@xilinx.com>
Acked-by: Naga Sureshkumar Relli <naga.sureshkumar.relli@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>