Revert "net: axienet: convert to phylink API"
commit b49b06d93560ffc98f7352c0bfcf92dadf630399 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

This reverts commit f5203a3d9b25fa28a40bfc27c05a4020c6430f06.
There are reported link issues with SGMII and 1000Base-x MLO_AN_PHY
mode. To fix it we are in discussion with mainline and Russell
suggested extension to phylink framework and modification to axieth
driver which is work in progress. For now revert phylink support
till all xilinx usecases are supported and functional in the
mainline tree.

Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
Signed-off-by: Swati Agarwal <swati.agarwal@xilinx.com>
Link: https://lkml.org/lkml/2020/1/27/603
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>