net: xilinx: axiethernet: Pass phy interface to of_phy_connect()
commit 315b4ba1c1e933511b610db0cfb440b572cf1826 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Pass phy interface to of_phy_connect() for all the modes (RGMII, GMII,
SGMII,and 1000BaseX). This change makes phy interface a configurable
DT property. It is required to bringup the axiethernet RGMII on kc705
which needs "rgmii-rxid" phy-interface to be passed to the Marvell
(M88E1111) phy driver.

Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
Signed-off-by: Swati Agarwal <swati.agarwal@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>