i2c: cadence: Clear HOLD bit before xfer_size register rolls over
commit  9f489509415e1c77668ab3b867f70eca1e74cfe1 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

On Xilinx zynq SOC if the delay between address register write and
control register write in cdns_mrecv function is more, the xfer size
register rolls over and controller is stuck. This is an IP bug and
is resolved in later versions of IP.

To avoid this scenario, disable the interrupts on the current processor
core between the two register writes and enable them later. This can
help achieve the timing constraint. Since the bug is only on zynq
platform, avoid doing this for other platforms.

Signed-off-by: Raviteja Narayanam <raviteja.narayanam@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>