usb: dwc3: otg: Enable OTG support in dwc3-core driver
commit  4699cb82557c4fcd9f9f8a08efd36fa604e3772a from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

The Linux v5.4 Kernel does not support OTG feature in the DWC3 USB driver.
It has some checks to set the mode to peripheral when the dr_mode is set
OTG in the device tree. To support the OTG mode for Xilinx platforms, the
check provided by the mainline kernel needs to be removed. This patch
removes the check-in dwc3 driver to make OTG work for Xilinx platforms.

Signed-off-by: Piyush Mehta <piyush.mehta@xilinx.com>
Acked-by: Manish Narani <manish.narani@xilinx.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>