firmware: zynqmp: Use DEVICE_ATTR_* to create sysfs entry
commit 315a742b1d9bfc8ff5428fd7997dfb8d4c623987 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

If sysfs entry is created using kobject, there are chances that,
kobject was created, and notified userspace that it was present and
then sometime after that if more attributes are added, which
userspace has no idea if they are there and such sysfs entry will be
probably missed.

So, use DEVICE_ATTR_* to create sysfs entry of last_reset_reason,
same as other sysfs entries are created in zynqmp firmware driver of
Linux v5.10.

Signed-off-by: Tejas Patel <tejas.patel@xilinx.com>
Signed-off-by: Rajan Vaja <rajan.vaja@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>