drm: xlnx: zynqmp_dpsub: Fix pixel clock issue
commit  147995f5c0dbdf8c8d245ece6d72cce7f61fdb5f from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

On boot-up DP display is not getting due to pixel clock issue,
The initial hardware state is unknown, and this makes sure that
the clock is disabled.

Ideally this sequence is not required but there is some issue (mostly
with ccf) which leads issue in DP clock(pclk), the value for Fractional-
control for the PLL is being set but the enable bit is not getting updated.
This is to have DP DRM working as of now, and latter will have proper fix.

Signed-off-by: Rohit Visavalia <rohit.visavalia@xilinx.com>
Link: https://lore.kernel.org/r/YFEV1ehZfivgBoKn@pendragon.ideasonboard.com
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>