remoteproc: zynqmp_r5: only request core nodes for zynqmp platform
commit  ad7fa5c7bfba04279265b49f9a53557972fc62ed from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

As the request call for Versal and ZynqMP have different assumptions,
making the same call works in most but not all cases. Namely core nodes.

On Versal, cores are individual nodes to be powered on and can belong to
different subsytems (LPD and FPD for example).

On ZynqMP this is not the same. On ZynqMP cores are themselves 'masters'
that can not be requested or released. But instead enable a xilpm client
(in this case Linux) to request or release a slave node like one of the
RPU's TCM banks.

Signed-off-by: Ben Levinsky <ben.levinsky@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>